refact(*): Cleaning CentOS 7 and Ubuntu 14.04 to 18.04, Ubuntu 20.04 included
Created the Jinja2 template file for the Debian-based Docker images,
at the moment this template is being used for Debian and Ubuntu.

CentOS 7 Docker file was cleaned up, in a future release CentOS 6 will
be included in this repository, just for the sake of testing with old
OS versions the Ansible roles/playbooks.