feat(automation): Including dynamic OpenJDK and pip selection
During Docker image build time, the openjdk available on the indexes
will be used. As long as they match the regex: OpenJDK 8 or 11.

When easy_install isn't find on the OS, the available pip (whether pip3 or pip2)
will be returned with the "install" parameter. i.e.: "/usr/sbin/pip?
install". This gives more freedom on the usage of pip3 or pip3, and
ignore OS versions where the easy_install isn't available.

From 2020 on, support for Python 2 is dropped. Therefore, easy_install
and some old tools will appear less and less on the OS. Thus, the
implementation of this dynamic handling was necessary.